Split components depending on compiler (#14)
Initial idea was to have common components and include files depending
on used compiler, i.e. ARMCC and GCC.
While this certainly looks better visually and in terms of maintenance
it does not work nice when implementation changes.

Single component version (per implementation) could work for components
with the same API but this breaks with
OS Interface where the APIs are different for ARMCC and GCC. It is hard
to justify a component version change
where there is no change in implementation file. This happens when
implementation for ARMCC changes, but remains
the same for GCC.

Although I don't like component duplication because the pdsc becomes
somewhat bloated, it seems like a right approach going forward.